corrected/updated external function names passed as arguments in pseudosoft.F90 and subgrids.F90

+++ code/source_f90/pseudosoft.F90
@@ -136,7 +136,7 @@ REAL(DP),DIMENSION(:),ALLOCATABLE :: pseudorho,potsave,potshort
 INTEGER,EXTERNAL :: isoutofbox
 INTEGER,EXTERNAL :: conv3to1
 INTEGER,EXTERNAL :: getnearestgridpoint
-INTEGER,EXTERNAL :: v_soft
+REAL(DP),EXTERNAL :: v_soft
 
 !--------------------------------------------------------------
